codes/requirements.txt
kafka-python
requests
SQLAlchemy
psycopg2-binary